favourites shows for signed-in user

+++ packages/client/src/components/ToolItem/ToolItem.component.js
@@ -3,8 +3,10 @@ import PropTypes from 'prop-types';
 import './ToolItem.style.css';
 import { Link } from 'react-router-dom';
 import getApiBaseUrl from '../../utils/getApiBaseURL';
+import { UserAuth } from '../../firebase/AuthContext';
 
 export const ToolItem = ({ tool }) => {
+  const { userId } = UserAuth();
   const title = tool.name;
   const { picture } = tool;
   const { id } = tool;
@@ -17,47 +19,50 @@ export const ToolItem = ({ tool }) => {
 
   const [isFavourite, setIsFavourite] = useState(false);
 
-  // using this user_id needed before Firebase implemented
-  const userId = 1;
-
   const handleChangeFavourite = () => {
-    if (isFavourite) {
-      fetch(`${getApiBaseUrl()}/api/favourites`, {
-        method: 'DELETE',
-        body: JSON.stringify({ user_id: userId, tool_id: id }),
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      }).then((result) => {
-        if (result.ok) {
-          setIsFavourite(false);
-        }
-      });
-    } else {
-      fetch(`${getApiBaseUrl()}/api/favourites`, {
-        method: 'POST',
-        body: JSON.stringify({ user_id: userId, tool_id: id }),
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      }).then((result) => {
-        if (result.ok) {
-          setIsFavourite(true);
-        }
-      });
+    if (userId) {
+      if (isFavourite) {
+        fetch(`${getApiBaseUrl()}/api/favourites`, {
+          method: 'DELETE',
+          body: JSON.stringify({ user_id: userId, tool_id: id }),
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        }).then((result) => {
+          if (result.ok) {
+            setIsFavourite(false);
+          }
+        });
+      } else {
+        fetch(`${getApiBaseUrl()}/api/favourites`, {
+          method: 'POST',
+          body: JSON.stringify({ user_id: userId, tool_id: id }),
+          headers: {
+            'Content-Type': 'application/json',
+          },
+        }).then((result) => {
+          if (result.ok) {
+            setIsFavourite(true);
+          }
+        });
+      }
     }
   };
 
   useEffect(() => {
-    fetch(`${getApiBaseUrl()}/api/favourites/${userId}`)
-      .then((response) => response.json())
-      .then((result) => {
-        const filteredResult = result.filter((item) => item.id === id);
-        if (filteredResult.length > 0) {
-          setIsFavourite(true);
-        }
-      });
-  }, [id]);
+    if (userId) {
+      fetch(`${getApiBaseUrl()}/api/favourites/${userId}`)
+        .then((response) => response.json())
+        .then((result) => {
+          const filteredResult = result.filter((item) => item.id === id);
+          if (filteredResult.length > 0) {
+            setIsFavourite(true);
+          }
+        });
+    } else {
+      setIsFavourite(false);
+    }
+  }, [userId, id]);
 
   const availableCategories = ['Innovation', 'Action', 'Energizer', 'Team'];
   const categoriesStreakOut = availableCategories.map((category) => {
